Fixes script on provisioning
This task is used on provisioning, and if it can't find any nginx (which
it won't on first provision, grep will throw an error which causes the
ansible script to fail. The standard error it throws is "" though, and
rc is 1. I only want errors to be thrown by the tasks if rc is not 0 (0
is non-error), and the stderr is not "". That way, other errors will get
triggered, but the issue of finding nothing won't break the script.